Unparsing ELF segment/section tables; .bss section
Segments that didn't occupy file memory were not initially included
in the list of sections since a section is a region of a file. Now
such segments (like .bss) are added to the section list with a size
of zero and whatever address appears in the section header table.


git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@810 a27c51de-2d2b-4d1b-a6e8-f7f87814f453

+++ src/frontend/ExecFormats/ExecGeneric.h
@@ -643,7 +643,7 @@ inline void host_to_le(unsigned h, uint32_t &n) {
 }
 inline void host_to_le(addr_t h, uint64_t &n) {
     assert(0==(h & ~0xfffffffffffffffful));
-    uint8_t hh = h;
+    uint64_t hh = h;
     n = ORDER_LSB==host_order() ? hh : swap_bytes(hh);
 }
 inline void host_to_le(int h, int8_t &n) {
@@ -708,7 +708,7 @@ inline void host_to_be(unsigned h, uint32_t &n) {
 }
 inline void host_to_be(addr_t h, uint64_t &n) {
     assert(0==(h & ~0xfffffffffffffffful));
-    uint8_t hh = h;
+    uint64_t hh = h;
     n = ORDER_MSB==host_order() ? hh : swap_bytes(hh);
 }
 inline void host_to_be(int h, int8_t &n) {
